Add files via upload
Sliding Login and Registration section in one page animating both section on click using only Html, Css and Javascript Download the files and it is ready to work. Responsive design in web view it is sliding horizontally and in mobile view login and registration sections are sliding vertically. Hope you like and this and I wish it would be useful to anyone out there. If you like it please follow @babayaga_coding on Instagram.